+++ src/lib/Errors.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 Angus.Fenying <fenying@litert.org>
+ * Copyright 2024 Angus.Fenying <fenying@litert.org>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,167 +14,305 @@
  * limitations under the License.
  */
 
-import * as $Exceptions from '@litert/exception';
+/**
+ * The error class for websocket.
+ */
+export abstract class OttoiaError extends Error {
 
-export const errorRegistry = $Exceptions.createExceptionRegistry({
-    'module': 'ottoia.litert.org',
-    'types': {
-        'public': {
-            'index': $Exceptions.createIncreaseCodeIndex(1)
-        }
+    public constructor(
+        name: string,
+        message: string,
+        public readonly context: Record<string, unknown>,
+        public readonly origin: unknown = null
+    ) {
+
+        super(message);
+        this.name = name;
     }
-});
+}
 
-export const E_PACKAGE_NOT_FOUND = errorRegistry.register({
-    name: 'package_not_found',
-    message: 'No such a package.',
-    metadata: {},
-    type: 'public'
-});
+export const E_PACKAGE_NOT_FOUND = class extends OttoiaError {
 
-export const E_NPM_ERROR = errorRegistry.register({
-    name: 'npm_error',
-    message: 'Failed with NPM operation.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_PACKAGE_NOT_RELEASED = errorRegistry.register({
-    name: 'package_not_released',
-    message: 'The package has not been published yet.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'package_not_found',
+            'No such a package.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_INVALID_PACKAGE = errorRegistry.register({
-    name: 'invalid_package',
-    message: 'The package.json is invalid.',
-    metadata: {},
-    type: 'public'
-});
+export const E_NPM_ERROR = class extends OttoiaError {
 
-export const E_INVALID_PATH = errorRegistry.register({
-    name: 'invalid_path',
-    message: 'The path could not be accessed.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_DUP_PACKAGE = errorRegistry.register({
-    name: 'dup_package',
-    message: 'The determined package already exists.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'npm_error',
+            'Failed with NPM operation.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_RELEASE_CONFIG_NOT_FOUND = errorRegistry.register({
-    name: 'release_config_not_found',
-    message: 'No such a release configuration.',
-    metadata: {},
-    type: 'public'
-});
+export const E_PACKAGE_NOT_RELEASED = class extends OttoiaError {
 
-export const E_COMMAND_NOT_FOUND = errorRegistry.register({
-    name: 'command_not_found',
-    message: 'No such a command.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_DUP_PACKAGE_ALIAS = errorRegistry.register({
-    name: 'dup_package_alias',
-    message: 'The determined package alias already exists.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'package_not_released',
+            'The package has not been published yet.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_RECURSIVE_DEP = errorRegistry.register({
-    name: 'recursive_dep',
-    message: 'Recursive dependency is forbidden.',
-    metadata: {},
-    type: 'public'
-});
+export const E_INVALID_PACKAGE = class extends OttoiaError {
 
-export const E_INVALID_TEMPLATE = errorRegistry.register({
-    name: 'invalid_template',
-    message: 'The template of package is invalid.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_INVALID_PACKAGE_NAME = errorRegistry.register({
-    name: 'invalid_package_name',
-    message: 'The name of package is invalid.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'invalid_package',
+            'The package.json is invalid.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_INVALID_JSON_FILE = errorRegistry.register({
-    name: 'invalid_json_file',
-    message: 'The JSON is malformed.',
-    metadata: {},
-    type: 'public'
-});
+export const E_INVALID_PATH = class extends OttoiaError {
 
-export const E_INVALID_ROOT_PACKAGE = errorRegistry.register({
-    name: 'invalid_root_package',
-    message: 'The package.json of root package is malformed.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_NO_ROOT_PACKAGE = errorRegistry.register({
-    name: 'no_root_package',
-    message: 'The root package not found.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'invalid_path',
+            'The path could not be accessed.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_SHELL_FAILED = errorRegistry.register({
-    name: 'shell_failed',
-    message: 'Failed to execute shell command.',
-    metadata: {},
-    type: 'public'
-});
+export const E_DUP_PACKAGE = class extends OttoiaError {
 
-export const E_EXISTING_PACKAGE_JSON = errorRegistry.register({
-    name: 'existing_package_json',
-    message: 'Can not initialize an existing package.json without "--yes" option.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_UNKNOWN_SUB_PACKAGE = errorRegistry.register({
-    name: 'unknown_sub_package',
-    message: 'The determined sub packages does not exist.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'dup_package',
+            'The determined package already exists.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_DEP_NOT_FOUND = errorRegistry.register({
-    name: 'dep_not_found',
-    message: 'No such a dependency.',
-    metadata: {},
-    type: 'public'
-});
+export const E_RELEASE_CONFIG_NOT_FOUND = class extends OttoiaError {
 
-export const E_DEP_INVALID = errorRegistry.register({
-    name: 'dep_invalid',
-    message: 'Can not resolve the dependency expression.',
-    metadata: {},
-    type: 'public'
-});
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
 
-export const E_DEP_NOT_LOCKED = errorRegistry.register({
-    name: 'dep_not_locked',
-    message: 'No such a dependency in the root package.json.',
-    metadata: {},
-    type: 'public'
-});
+        super(
+            'release_config_not_found',
+            'No such a release configuration.',
+            ctx,
+            origin
+        );
+    }
+};
 
-export const E_PRIVATE_DEPENDENCY = errorRegistry.register({
-    name: 'private_dependency',
-    message: 'Can not depend on a private sub package.',
-    metadata: {},
-    type: 'public'
-});
+export const E_COMMAND_NOT_FOUND = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'command_not_found',
+            'No such a command.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_DUP_PACKAGE_ALIAS = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'dup_package_alias',
+            'The determined package alias already exists.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_RECURSIVE_DEP = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'recursive_dep',
+            'Recursive dependency is forbidden.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_INVALID_TEMPLATE = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'invalid_template',
+            'The template of package is invalid.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_INVALID_PACKAGE_NAME = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'invalid_package_name',
+            'The name of package is invalid.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_INVALID_JSON_FILE = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'invalid_json_file',
+            'The JSON is malformed.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_INVALID_ROOT_PACKAGE = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'invalid_root_package',
+            'The package.json of root package is malformed.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_NO_ROOT_PACKAGE = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'no_root_package',
+            'The root package not found.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_SHELL_FAILED = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'shell_failed',
+            'Failed to execute shell command.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_EXISTING_PACKAGE_JSON = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'existing_package_json',
+            'Can not initialize an existing package.json without "--yes" option.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_UNKNOWN_SUB_PACKAGE = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'unknown_sub_package',
+            'The determined sub packages does not exist.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_DEP_NOT_FOUND = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'dep_not_found',
+            'No such a dependency.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_DEP_INVALID = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'dep_invalid',
+            'Can not resolve the dependency expression.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_DEP_NOT_LOCKED = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'dep_not_locked',
+            'No such a dependency in the root package.json.',
+            ctx,
+            origin
+        );
+    }
+};
+
+export const E_PRIVATE_DEPENDENCY = class extends OttoiaError {
+
+    public constructor(ctx: Record<string, unknown> = {}, origin: unknown = null) {
+
+        super(
+            'private_dependency',
+            'Can not depend on a private sub package.',
+            ctx,
+            origin
+        );
+    }
+};